Improve behavior of post and postPower (#1447)
* Improve behavior of post and postPower

The `post()` and `postPower()` functions of various structures are now
called by the game loop, not by 250ms delay setTimeout invocations.
This means that they cannot be delayed by event queue congestion.
These functions are invoked at the same frequency as fastLoop, but the
actual call site is placed between midLoop and longLoop.

Remove some incorrect code from postPower() for telemetry beacons.

Use a more precise method to detect when the longLoop should redraw the
city tab in response to an updated value for city categorization.

Don't attempt to redraw the power grid in longLoop when the power grid
isn't visible and preload tab content is disabled.

* Disable reset structures that lose power later